requirements.txt
flask
pyYAML
prometheus-flask-exporter
